+++ lib/streaming.mli
@@ -37,6 +37,7 @@ type stream_acc =
   ; block_tool_ids : (int, string) Hashtbl.t
   ; block_tool_names : (int, string) Hashtbl.t
   ; block_thinking_signatures : (int, Buffer.t) Hashtbl.t
+  ; block_reasoning_details : (int, Types.reasoning_detail list ref) Hashtbl.t
   ; block_media_types : (int, string) Hashtbl.t
   ; block_media_sources : (int, Types.media_source_kind) Hashtbl.t
   }